Ajustado a Integer el acelerómetro y gisroscopio del reloj

+++ app/src/main/java/com/hdrescuer/hdrescuer/data/GlobalMonitoringViewModel.java
@@ -24,15 +24,15 @@ public class GlobalMonitoringViewModel extends AndroidViewModel implements ViewM
     private MutableLiveData<Float> hb;
     private MutableLiveData<Float> hrppgraw;
     private MutableLiveData<Integer> step;
-    private MutableLiveData<Float> accx;
-    private MutableLiveData<Float> accy;
-    private MutableLiveData<Float> accz;
-    private MutableLiveData<Float> acclx;
-    private MutableLiveData<Float> accly;
-    private MutableLiveData<Float> acclz;
-    private MutableLiveData<Float> girx;
-    private MutableLiveData<Float> giry;
-    private MutableLiveData<Float> girz;
+    private MutableLiveData<Integer> accx;
+    private MutableLiveData<Integer> accy;
+    private MutableLiveData<Integer> accz;
+    private MutableLiveData<Integer> acclx;
+    private MutableLiveData<Integer> accly;
+    private MutableLiveData<Integer> acclz;
+    private MutableLiveData<Integer> girx;
+    private MutableLiveData<Integer> giry;
+    private MutableLiveData<Integer> girz;
 
     //Atributos de los sensores de la Empática
     private MutableLiveData<Float> battery;
@@ -104,37 +104,37 @@ public class GlobalMonitoringViewModel extends AndroidViewModel implements ViewM
         return step;
     }
 
-    public MutableLiveData<Float> getAccx() { return accx; }
+    public MutableLiveData<Integer> getAccx() { return accx; }
 
-    public MutableLiveData<Float> getAccy() {
+    public MutableLiveData<Integer> getAccy() {
         return accy;
     }
 
-    public MutableLiveData<Float> getAccz() {
+    public MutableLiveData<Integer> getAccz() {
         return accz;
     }
 
-    public MutableLiveData<Float> getAcclx() {
+    public MutableLiveData<Integer> getAcclx() {
         return acclx;
     }
 
-    public MutableLiveData<Float> getAccly() {
+    public MutableLiveData<Integer> getAccly() {
         return accly;
     }
 
-    public MutableLiveData<Float> getAcclz() {
+    public MutableLiveData<Integer> getAcclz() {
         return acclz;
     }
 
-    public MutableLiveData<Float> getGirx() {
+    public MutableLiveData<Integer> getGirx() {
         return girx;
     }
 
-    public MutableLiveData<Float> getGiry() {
+    public MutableLiveData<Integer> getGiry() {
         return giry;
     }
 
-    public MutableLiveData<Float> getGirz() {
+    public MutableLiveData<Integer> getGirz() {
         return girz;
     }
 
@@ -205,31 +205,31 @@ public class GlobalMonitoringViewModel extends AndroidViewModel implements ViewM
 
     public void setStep(Integer step){ this.step.postValue(step);}
 
-    public void setAccx(Float accx) { this.accx.postValue(accx);   }
+    public void setAccx(Integer accx) { this.accx.postValue(accx);   }
 
-    public void setAccy(Float accy) { this.accy.postValue(accy);   }
+    public void setAccy(Integer accy) { this.accy.postValue(accy);   }
 
-    public void setAccz(Float accz) { this.accz.postValue(accz);   }
+    public void setAccz(Integer accz) { this.accz.postValue(accz);   }
 
-    public void setAcclx(Float acclx) {
+    public void setAcclx(Integer acclx) {
         this.acclx.postValue(acclx);
     }
 
-    public void setAccly(Float accly) {
+    public void setAccly(Integer accly) {
         this.accly.postValue(accly);
     }
 
-    public void setAcclz(Float acclz) {
+    public void setAcclz(Integer acclz) {
         this.acclz.postValue(acclz);
     }
 
-    public void setGirx(Float girx) { this.girx.postValue(girx);  }
+    public void setGirx(Integer girx) { this.girx.postValue(girx);  }
 
-    public void setGiry(Float giry) {
+    public void setGiry(Integer giry) {
         this.giry.postValue(giry);
     }
 
-    public void setGirz(Float girz) {
+    public void setGirz(Integer girz) {
         this.girz.postValue(girz);
     }
